requirements.txt
flask==3.0.3
yfinance==0.2.38
gunicorn==22.0.0
Werkzeug==3.0.3
requests==2.32.3
pandas==2.2.2
numpy==1.26.4
python-dateutil==2.9.0
pytz==2024.1
psycopg2-binary==2.9.9
scipy==1.13.1
statsmodels==0.14.2
bcrypt==4.1.3